# Optimization: Reduce ImporBuf ArrayBuffer allocations