Better at http benchmark:
```
~/src/deno> ./tools/http_benchmark.py out/before/deno
http_benchmark testing DENO.
Listening on 127.0.0.1:4544
third_party/wrk/mac/wrk -d 10s http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   790.69us    1.22ms  23.65ms   95.58%
    Req/Sec     8.25k     1.10k    9.80k    88.50%
  164125 requests in 10.00s, 7.98MB read
Requests/sec:  16407.65
Transfer/sec:    817.18KB

~/src/deno> ./tools/http_benchmark.py out/after/deno
http_benchmark testing DENO.
Listening on 127.0.0.1:4544
third_party/wrk/mac/wrk -d 10s http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   640.84us    0.91ms  24.95ms   97.02%
    Req/Sec     9.03k     1.05k   10.16k    94.00%
  179642 requests in 10.00s, 8.74MB read
Requests/sec:  17962.53
Transfer/sec:      0.87MB
```
But worse at throughput benchmark (???)
```
~/src/deno> ./tools/throughput_benchmark.py out/before/deno 1000
listening on 127.0.0.1:4544
head -c 1048576000 /dev/zero | nc 127.0.0.1 4544
17.5132410526 seconds

~/src/deno> ./tools/throughput_benchmark.py out/after/deno 1000
listening on 127.0.0.1:4544
head -c 1048576000 /dev/zero | nc 127.0.0.1 4544
21.8031110764 seconds
```